style: work in progress on header styling
