+++ src/components/Auth/SignInWithOAuth/SignInWithOAuth.jsx
@@ -1,7 +1,6 @@
 import { OAUTH, OAUTH_PROVIDERS } from "../../../constants/constants";
 import useAuth from "../../../hooks/useAuth";
 
-
 const SignInWithOAuth = () => {
   const { handleAuth } = useAuth(OAUTH);
   return (
